Use iostream, not iostream.h Include assert.h in Cassowary.h, not elsewhere Include auto_ptr.h where needed, and have it include <memory> for VC++ Define operator !=, operator>, etc., that I was getting for free   with egcs from function.h, since that isn't standard Don't use variable-sized arrays (is g++ extension) Use DBL_MAX, not MAX_DOUBLE Assume auto_ptr doesn't have .reset -- use another auto_ptr +   assignment Rename ClSymbolicWeight::{plus,minus} to ::{add,subtract} Added bool return types to operator{==,<,etc.} for ClSymbolicWeight Don't name unused variables (e.g., &error) in test programs Return a value from main()-s Update makefile for some nt stuff, including ntsources target to make   symlinks of *.cpp filenames Other random fixes for VC++:   o turn off extra long identifier truncation warning   o #define __FUNCTION__ to "WIN" (__FUNCTION__ is g++ extension)

+++ c++/ClVariable.h
@@ -11,9 +11,8 @@
 #ifndef ClVariable_H
 #define ClVariable_H
 
-#include <iostream.h>
+#include <iostream>
 #include <stdio.h>
-#include <assert.h>
 #include "Cassowary.h"
 #include "ClAbstractVariable.h"
 
@@ -81,7 +80,7 @@ private:
 // Compare two double-s approximately, since equality is no good
 inline bool clApprox(double a, double b)
 {
-  float epsilon = 1.0e-8;
+  double epsilon = 1.0e-8;
   if (a == 0.0) {
     return (fabs(b) < epsilon);
   } else if (b == 0.0) {
